requirements.txt
# Core data processing
pandas==1.5.3
numpy==1.24.4
pyarrow==21.0.0

# NFL data
nfl_data_py==0.3.3

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0
joblib>=1.3.0

# Web Dashboard
streamlit>=1.28.0
plotly>=5.17.0

# Development/Testing
pytest>=7.4.0
